server/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
flask_cors=
dotenv=
requests=
google.generativeai